Update -Change raw score to decimal type in case raw score two digit -Fixed bug in save edit score puchase not set action -Add Predicate Builder class -Change then by desc in period to asc -Change to using entity vendor transaction

+++ EVF.Evaluation.Bll/Models/EvaluationLogViewModel.cs
@@ -31,7 +31,7 @@ namespace EVF.Evaluation.Bll.Models
         [MaxLength(255)]
         public string Reason { get; set; }
         public int MaxScore { get; set; }
-        public int RawScore { get; set; }
+        public double RawScore { get; set; }
     }
 
 }
